requirements.txt
requests
beautifulsoup4
ctransformers==0.2.5
sentence-transformers==2.2.2
langchain==0.0.225
orjson
flask
ipykernel
python-dotenv
huggingface-hub==0.14.1
qdrant-client
-e . 